requirements.txt
Django==4.2.4
Pillow==10.0.0
scipy==1.10.0
